+++ packages/hasura-storage-js/src/utils/types.ts
@@ -48,10 +48,25 @@ export type StorageUploadResponse =
   | { fileMetadata: FileResponse; error: null }
   | { fileMetadata: null; error: Error }
 
-export interface StorageGetUrlParams {
+export interface StorageImageTransformationParams {
+  /** Image width, in pixels */
+  width?: number
+  /** Image height, in pixels */
+  height?: number
+  /** Image quality, between 1 and 100, 100 being the best quality */
+  quality?: number
+  /** Image blur, between 0 and 100 */
+  blur?: number
+  // TODO not implemented yet in hasura-storage
+  /** Image radius */
+  // radius?: number
+}
+export interface StorageGetUrlParams extends StorageImageTransformationParams {
   fileId: string
 }
 
+// TODO not implemented yet in hasura-storage
+// export interface StorageGetPresignedUrlParams extends StorageImageTransformationParams {
 export interface StorageGetPresignedUrlParams {
   fileId: string
 }
@@ -89,6 +104,8 @@ export type ApiUploadResponse =
   | { fileMetadata: FileResponse; error: null }
   | { fileMetadata: null; error: Error }
 
+// TODO not implemented yet in hasura-storage
+// export interface ApiGetPresignedUrlParams extends StorageImageTransformationParams {
 export interface ApiGetPresignedUrlParams {
   fileId: string
 }